# docs/api-contracts/openapi-v1-baseline.yaml
...
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
...
paths:
...
  /api/v1/organizations/{id}:
    get:
      tags:
      - Organizations
      summary: Get an Organization by ID
      description: Fetch a single Organization scoped to a specific Tenant. The tenantId query parameter is required. If the (tenantId, id) pair does not match any Organization, a 404 is returned.
      operationId: getOrganization
      parameters:
      - name: id
        in: path
        description: Organization UUID
        required: true
        schema:
          type: string
          format: uuid
      - name: tenantId
        in: query
        description: Tenant UUID (scope)
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Organization found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '400':
          description: Bad Request - missing required tenantId query parameter
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Organization Not Found - no Organization with the given id exists under the given tenant
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    put:
      tags:
      - Organizations
      summary: Update an Organization's name and description
      description: Updates the mutable fields of an Organization. The tenant relationship and status are NOT changed. If the new name conflicts with another organization under the same tenant, a 409 Conflict is returned.
      operationId: updateOrganization
      parameters:
      - name: id
        in: path
        description: Organization UUID
        required: true
        schema:
          type: string
          format: uuid
      - name: tenantId
        in: query
        description: Tenant UUID (scope)
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationRequest'
        required: true
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '400':
          description: Bad Request - request body failed validation or tenantId missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Organization Not Found - no Organization with the given id exists under this tenant
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Organization Already Exists - another organization under the same tenant already uses the new name
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/v1/organizations:
    get:
      tags:
      - Organizations
      summary: List Organizations for a Tenant
      description: Returns all Organizations belonging to the specified Tenant. The tenantId query parameter is required. If the tenant has no organizations, an empty array is returned.
      operationId: listOrganizations
      parameters:
      - name: tenantId
        in: query
        description: Tenant UUID (scope)
        required: true
        schema:
          type: string
          format: uuid
      - name: params
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/PageRequestParams'
      responses:
        '200':
          description: Organization list (possibly empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationResponse'
        '400':
          description: Bad Request - missing required tenantId query parameter
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
    post:
      tags:
      - Organizations
      summary: Create a new Organization
      description: Registers a new Organization under an existing Tenant. The Organization is created with status ACTIVE. The (tenantId, name) pair must be unique; otherwise a 409 is returned.
      operationId: createOrganization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
        required: true
      responses:
        '201':
          description: Organization created successfully
          headers:
            Location:
              description: URI of the newly created Organization resource
              style: simple
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '400':
          description: Bad Request - request body failed validation (missing tenantId, blank name, name or description too long)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Tenant Not Found - the referenced tenantId does not exist
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Organization Already Exists - an Organization with the same name already exists under the same Tenant
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/v1/organizations/{id}/deactivate:
    patch:
      tags:
      - Organizations
      summary: Deactivate an Organization
      description: 'Sets the Organization status to INACTIVE. Idempotent: deactivating an already-INACTIVE organization is a no-op that returns the current state.'
      operationId: deactivateOrganization
      parameters:
      - name: id
        in: path
        description: Organization UUID
        required: true
        schema:
          type: string
          format: uuid
      - name: tenantId
        in: query
        description: Tenant UUID (scope)
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Organization deactivated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '400':
          description: Bad Request - missing required tenantId query parameter
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Organization Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/organizations/{id}/archive:
    patch:
      tags:
      - Organizations
      summary: Archive an Organization (soft delete)
      description: 'Marks the Organization as ARCHIVED without removing it from the database. Idempotent: archiving an already-archived organization is a no-op that returns the current state. Use this instead of a DELETE endpoint to preserve audit history.'
      operationId: archiveOrganization
      parameters:
      - name: id
        in: path
        description: Organization UUID
        required: true
        schema:
          type: string
          format: uuid
      - name: tenantId
        in: query
        description: Tenant UUID (scope)
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Organization archived successfully (or was already archived)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '400':
          description: Bad Request - missing required tenantId query parameter
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Organization Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/organizations/{id}/activate:
    patch:
      tags:
      - Organizations
      summary: Activate an Organization
      description: 'Sets the Organization status to ACTIVE. Idempotent: activating an already-ACTIVE organization is a no-op that returns the current state.'
      operationId: activateOrganization
      parameters:
      - name: id
        in: path
        description: Organization UUID
        required: true
        schema:
          type: string
          format: uuid
      - name: tenantId
        in: query
        description: Tenant UUID (scope)
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Organization activated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '400':
          description: Bad Request - missing required tenantId query parameter
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Organization Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /api/v1/access/roles/{roleId}/deactivate:
    patch:
      tags:
      - role-controller
      operationId: deactivate
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
  /api/v1/access/roles/{roleId}/archive:
    patch:
      tags:
      - role-controller
      operationId: archive
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
  /api/v1/access/roles/{roleId}/activate:
    patch:
      tags:
      - role-controller
      operationId: activate
      parameters:
      - name: tenantId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
  /api/v1/access/capabilities/{capabilityId}/deactivate:
    patch:
      tags:
      - capability-controller
      operationId: deactivate_1
      parameters:
      - name: capabilityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityResponse'
  /api/v1/access/capabilities/{capabilityId}/activate:
    patch:
      tags:
      - capability-controller
      operationId: activate_1
      parameters:
      - name: capabilityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityResponse'
...
components:
  schemas:
...
    ApiErrorResponse:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldValidationError'
    FieldValidationError:
      type: object
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 0
        description:
          type: string
          maxLength: 1000
          minLength: 0
      required:
      - name
    OrganizationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - ARCHIVED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...